requirements.txt
appnope==0.1.4; platform_system == "Darwin" and python_version >= "3.8"
asttokens==2.4.1; python_version >= "3.10"
beautifulsoup4==4.12.3; python_full_version >= "3.6.0"
bibtexparser==1.4.1
cachetools==5.3.3; python_version >= "3.7"
certifi==2024.2.2; python_version >= "3.8"
cffi==1.16.0; implementation_name == "pypy" and python_version >= "3.8"
charset-normalizer==3.3.2; python_full_version >= "3.7.0" and python_version >= "3.8"
colorama==0.4.6; python_version >= "3.10" and python_full_version < "3.0.0" and sys_platform == "win32" or sys_platform == "win32" and python_version >= "3.10" and python_full_version >= "3.7.0"
comm==0.2.2; python_version >= "3.8"
debugpy==1.8.1; python_version >= "3.8"
decorator==5.1.1; python_version >= "3.10"
doi2bibtex==0.2.0
exceptiongroup==1.2.1; python_version < "3.11" and python_version >= "3.10"
executing==2.0.1; python_version >= "3.10"
google-api-core==2.19.0; python_version >= "3.7"
google-api-python-client==2.130.0; python_version >= "3.7"
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0; python_version >= "3.6"
google-auth==2.29.0; python_version >= "3.7"
googleapis-common-protos==1.63.0; python_version >= "3.7"
httplib2==0.22.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.7"
idna==3.7; python_version >= "3.8"
ipykernel==6.29.4; python_version >= "3.8"
ipython==8.24.0; python_version >= "3.10"
isbnlib==3.10.14
jedi==0.19.1; python_version >= "3.10"
jupyter-client==8.6.2; python_version >= "3.8"
jupyter-core==5.7.2; python_version >= "3.8"
markdown-it-py==3.0.0; python_version >= "3.8" and python_full_version >= "3.7.0"
matplotlib-inline==0.1.7; python_version >= "3.10"
mdurl==0.1.2; python_version >= "3.8" and python_full_version >= "3.7.0"
nest-asyncio==1.6.0; python_version >= "3.8"
oauthlib==3.2.2; python_version >= "3.6"
packaging==24.0; python_version >= "3.8"
parso==0.8.4; python_version >= "3.10"
pexpect==4.9.0; sys_platform != "win32" and sys_platform != "emscripten" and python_version >= "3.10"
platformdirs==4.2.2; python_version >= "3.8"
prompt-toolkit==3.0.43; python_full_version >= "3.7.0" and python_version >= "3.10"
proto-plus==1.23.0; python_version >= "3.7"
protobuf==4.25.3; python_version >= "3.8"
psutil==5.9.8; python_version >= "3.8" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.8"
ptyprocess==0.7.0; sys_platform != "win32" and sys_platform != "emscripten" and python_version >= "3.10"
pure-eval==0.2.2; python_version >= "3.10"
pyasn1-modules==0.4.0; python_version >= "3.8"
pyasn1==0.6.0; python_version >= "3.8" and python_version < "4"
pycparser==2.22; implementation_name == "pypy" and python_version >= "3.8"
pygments==2.18.0; python_version >= "3.10" and python_full_version >= "3.7.0"
pylatexenc==2.10
pyparsing==3.1.2; python_full_version >= "3.6.8" and python_version >= "3.7"
python-dateutil==2.9.0.post0; python_version >= "3.8" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.8"
python-dotenv==1.0.1; python_version >= "3.8"
pywin32==306; sys_platform == "win32" and platform_python_implementation != "PyPy" and python_version >= "3.8"
pyyaml==6.0.1; python_version >= "3.6"
pyzmq==26.0.3; python_version >= "3.8"
requests-oauthlib==2.0.0; python_version >= "3.6"
requests==2.32.2; python_version >= "3.8"
rich==13.7.1; python_full_version >= "3.7.0"
rsa==4.9; python_version >= "3.7" and python_version < "4"
six==1.16.0; python_version >= "3.10" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.10"
soupsieve==2.5; python_version >= "3.8" and python_full_version >= "3.6.0"
stack-data==0.6.3; python_version >= "3.10"
tornado==6.4; python_version >= "3.8"
traitlets==5.14.3; python_version >= "3.10"
typing-extensions==4.12.0; python_version < "3.12" and python_version >= "3.10"
unidecode==1.3.8; python_version >= "3.5"
uritemplate==4.1.1; python_version >= "3.7"
urllib3==2.2.1; python_version >= "3.8"
wcwidth==0.2.13; python_full_version >= "3.7.0" and python_version >= "3.10"
